Work with new ogre2 debs

Approved-by: Louise Poubel <lupoubel@hotmail.com>
Approved-by: Steven Peters <scpeters@osrfoundation.org>

+++ ogre2/src/Ogre2RenderTarget.cc
@@ -17,7 +17,7 @@
 
 // leave this out of OgreIncludes as it conflicts with other files requiring
 // gl.h
-#include <OGRE/RenderSystems/GL3Plus/OgreGL3PlusFBORenderTexture.h>
+#include <OgreGL3PlusFBORenderTexture.h>
 
 #include <ignition/common/Console.hh>
 
@@ -295,15 +295,15 @@ void Ogre2RenderTexture::BuildTarget()
 }
 
 //////////////////////////////////////////////////
-GLuint Ogre2RenderTexture::GLId() const
+unsigned int Ogre2RenderTexture::GLId() const
 {
   if (!this->ogreTexture)
-    return GLuint(0);
+    return 0;
 
   GLuint texId;
   this->ogreTexture->getCustomAttribute("GLID", &texId);
 
-  return texId;
+  return static_cast<unsigned int>(texId);
 }
 
 //////////////////////////////////////////////////
